#20986 Migrate changes to existing CaffeineCache. Expirable interface. Make new class serializable
