+++ dotCMS/src/main/java/com/dotcms/business/Expirable.java
@@ -1,0 +1,5 @@
+package com.dotcms.business;
+
+public interface Expirable {
+    long getTtl();
+}
